Version 1.1
It's a script to insert a "Last Change" comment field in your source file.
You can modify the default behavior of this plug-in.
To add the support of new file type
put this to your vimrc
au BufWritePre *.cpp        call TimeStamp('//')
if the comment block in that language has both the begin and end symbol,
just enjoy this
au BufWritePre *.php          call TimeStamp('<?php //', '?>')